Use sinus instead of sigmoid
