+++ config/fre22_task_mapping_mini.yaml
@@ -1,11 +1,11 @@
 row_length: 5
 rows_count: 6
 row_segments:
-  - scurved
-row_segment_scurved_offset_min: 0.2
-row_segment_scurved_offset_max: 0.5
-row_segment_scurved_length_min: 5.0
-row_segment_scurved_length_max: 5.0
+  - sincurved
+row_segment_sincurved_offset_min: 0.2
+row_segment_sincurved_offset_max: 0.5
+row_segment_sincurved_length_min: 5.0
+row_segment_sincurved_length_max: 5.0
 hole_prob: [0.08, 0.08, 0.05, 0.02, 0.0, 0.0]
 hole_size_max: [8, 8, 6, 4, 0, 0]
 ground_resolution: 0.1
