+++ src/virtual_maize_field/world_generator/segments/__init__.py
@@ -1,13 +1,13 @@
 from .base import BaseSegment
 from .curve import CurvedSegment
 from .island import IslandSegment
-from .scurve import SCurvedSegment
+from .sincurve import SinCurvedSegment
 from .straight import StraightSegment
 
 __all__ = [
     "BaseSegment",
     "CurvedSegment",
     "IslandSegment",
-    "SCurvedSegment",
+    "SinCurvedSegment",
     "StraightSegment",
 ]
